api/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
celery==5.4.0
redis==5.0.7
boto3==1.34.162
requests==2.32.3
python-dotenv==1.0.1
loguru==0.7.2
